% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/at_risk.R
\name{at_risk}
\alias{at_risk}
\title{Calculate size of risk set at time x based on a vector of times t}
\usage{
at_risk(t, W, HR = NULL)
}
\arguments{
\item{t}{Value at which the size of the risk set is to be calculated.}

\item{W}{Numeric vector of observed covariate values (including censored opens).}

\item{HR}{(Optional) Numeric vector of hazard ratios (HR) corresponding to values in \code{W}.}
}
\value{
A scalar
}
\description{
Calculates the size of the risk set at specified times. Alternatively, if \code{HR} is not \code{NULL} then it sums up HR of at-risk people (for use with \code{breslow_estimator()}).
}
