% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breslow_estimator.R
\name{breslow_estimator}
\alias{breslow_estimator}
\title{Breslow estimator of baseline survival}
\usage{
breslow_estimator(time, event, hr, data)
}
\arguments{
\item{time}{String column name for the observed times.}

\item{event}{String column name for the censoring indicator of the covariate.}

\item{hr}{String column name for the hazard ratios calculated from a \code{coxph} model fit.}

\item{data}{Dataframe containing columns \code{time}, \code{event}, and \code{hr}.}
}
\value{
A list with the following three elements:
\item{times}{a vector of the unique observed failure times}
\item{basesurv}{baseline survival estimates}
\item{basehaz}{baseline cumulative hazard estimates}
}
\description{
Estimates the baseline survival function from a Cox proportional hazards model following Breslow's estimator.
}
