% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condl_mean_impute.R
\name{condl_mean_impute}
\alias{condl_mean_impute}
\title{Conditional mean single imputation}
\usage{
condl_mean_impute(
  fit,
  obs,
  delta,
  covar = NULL,
  data,
  approx_beyond = "expo",
  est_S0 = "breslow"
)
}
\arguments{
\item{fit}{A \code{coxph} or \code{survfit} imputation model object.}

\item{obs}{String column name for the censored covariate.}

\item{delta}{String column name for the censoring indicator of the covariate.}

\item{data}{Datafrane containing columns \code{obs}, \code{delta}, and (if provided) \code{addl_covar}.}

\item{approx_beyond}{Choice of approximation used to extrapolate the survival function beyond the last observed covariate value. Default is \code{"expo"} for the exponential approximation. Other choices include \code{"zero"} or \code{"carryforward"}.}

\item{addl_covar}{(Optional) string or vector of strings for the additional fully-observed covariates. Default is \code{NULL}.}
}
\value{
A dataframe augmenting \code{data} with a column of imputed covariate values called \code{imp}.
}
\description{
Imputes censored covariates with their conditional mean given censored value and additional covariates (where supplied).
}
