% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condl_mean_impute_bootstrap.R
\name{condl_mean_impute_bootstrap}
\alias{condl_mean_impute_bootstrap}
\title{Conditional mean imputation with bootstrap}
\usage{
condl_mean_impute_bootstrap(
  fit,
  obs,
  event,
  addl_covar = NULL,
  data,
  approx_beyond = "expo",
  M
)
}
\arguments{
\item{fit}{A \code{coxph} or \code{survfit} imputation model object.}

\item{obs}{String column name for the censored covariate.}

\item{event}{String column name for the censoring indicator of the covariate.}

\item{addl_covar}{(Optional) string or vector of strings for the additional fully-observed covariates. Default is \code{NULL}.}

\item{data}{Datafrane containing columns \code{obs}, \code{event}, and (if provided) \code{addl_covar}.}

\item{approx_beyond}{Choice of approximation used to extrapolate the survival function beyond the last observed covariate value. Default is \code{"expo"} for the exponential approximation. Other choices include \code{"zero"} or \code{"carryforward"}.}

\item{M}{number of bootstrap samples to be taken from \code{data}}
}
\value{
A list of \code{M} dataframes, each of which is sampled with replacement from \code{data} and then augmented with a column of imputed covariate values called \code{imp}.
}
\description{
Imputes censored covariates with their conditional mean given censored value and additional covariates (where supplied).
}
