% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate.data}
\alias{generate.data}
\title{Generate data for linear regression with censored covariate}
\usage{
generate.data(
  n,
  n.sims = 1,
  beta0,
  betaX,
  betaZ = NULL,
  sigma = 1,
  x = NULL,
  z = NULL,
  xshape = 0.75,
  xscale = 0.25,
  c.lower = 0,
  c.upper = 1
)
}
\arguments{
\item{n}{sample size for single simulation}

\item{n.sims}{number of simulations}

\item{beta0}{intercept}

\item{betaX}{coefficient for covariate x}

\item{betaZ}{coefficient for covariate(s) z. If betaZ NULL, then it is assumed that the model does not include z.}

\item{sigma}{sd for random error in outcome model}

\item{x}{the censored covariate. If x NULL, then x is generated from Weibull(xshape, xscale).}

\item{z}{the observed covariate(s). If betaZ is provided and z NULL, then z is generated from N_p(0, I).}

\item{xshape}{shape parameter for simulation of x (if x NULL)}

\item{xscale}{scale parameter for simulation of x (if x NULL)}

\item{c.lower}{minimum for simulation of c}

\item{c.upper}{maximum for simulation of c}
}
\value{

}
\description{
Generates data.frame pertaining to the linear model, Y = beta0 + X betaX + Z betaZ + epsilon
and a censored observation of the covariate X by censoring variable C
}
