% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate data for linear regression with censored covariate}
\usage{
generate_data(
  n,
  n.sims = 1,
  beta0,
  betaX,
  betaZ = NULL,
  sigma = 1,
  x = NULL,
  z = NULL,
  xshape = 0.75,
  xscale = 0.25,
  c.lower = 0,
  c.upper = 1
)
}
\arguments{
\item{n}{sample size for single simulation}

\item{n.sims}{number of simulations}

\item{beta0}{intercept}

\item{betaX}{coefficient for covariate \code{x}}

\item{betaZ}{coefficient for covariate(s) \code{z}. If \code{betaZ} NULL, then it is assumed that the model does not include \code{z}.}

\item{sigma}{sd for random error in outcome model}

\item{x}{censored covariate. If \code{x} NULL, then \code{x} is generated from Weibull(\code{xshape}, \code{xscale}).}

\item{z}{matrix of observed covariate(s). If \code{betaZ} is provided and \code{z} NULL, then \code{z} is generated from N_p(0, I).}

\item{xshape}{shape parameter for simulation of \code{x} (if \code{x} NULL)}

\item{xscale}{scale parameter for simulation of \code{x} (if \code{x} NULL)}

\item{c.lower}{minimum for simulation of \code{c}}

\item{c.upper}{maximum for simulation of \code{c}}
}
\value{
a data.frame with the following elements:
\itemize{
\item{subj.id} subject id
\item{sim.id} simulation id
\item{y} response
\item{x} true value of the covariate subject to censoring
\item{t} minimum of covariate \code{x} and censoring variable \code{c}
\item{delta} 1 if \code{x} <= \code{c}; 0 otherwise
\item{z} observed covariates
}
}
\description{
Generates data.frame pertaining to the linear model, Y = beta0 + X betaX + Z betaZ + epsilon
and a censored observation of the covariate X by censoring variable C
}
