% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_censored_surv.R
\name{impute_censored_surv}
\alias{impute_censored_surv}
\title{Impute survival at censored time}
\usage{
impute_censored_surv(at_time, time, event, surv, data)
}
\arguments{
\item{at_time}{A scalar time}

\item{time}{String column name for the censored covariate.}

\item{event}{String column name for the censoring indicator of the covariate.}

\item{surv}{String column name for the survival estimate.}

\item{data}{Datafrane containing columns \code{obs}, \code{event}, and \code{surv}.}
}
\value{
Scalar survival estimate for value \code{at_time}.
}
\description{
Imputes survival at censored covariate with mean of uncensored covariates before and after.
}
