% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breslow_estimator.R
\name{breslow_estimator}
\alias{breslow_estimator}
\title{Breslow estimator of baseline survival}
\usage{
breslow_estimator(x = NULL, time, event, hr, data)
}
\arguments{
\item{x}{(Optional) Value at which the baseline survival function is to be calculated. If \code{x = NULL} (the default), values at all unique event times is returned.}

\item{time}{Column name of observed predictor values (including censored opens).}

\item{event}{Column name of censoring indicators. Note that \code{Delta = 0} is interpreted as a censored observation.}

\item{hr}{Column name of hazard ratios (HR) from a Cox proportional hazards model.}

\item{data}{Dataframe or named matrix containing columns \code{W}, \code{Delta}, and \code{HR}.}
}
\value{
If \code{x} is not \code{NULL}, then a scalar value for the baseline survival function at value \code{x}. Otherwise, a list with the following three elements is returned:
\item{times}{a vector of the unique observed failure times}
\item{basesurv}{baseline survival estimates}
\item{basehaz}{baseline cumulative hazard estimates}
}
\description{
Estimates the baseline survival function from a Cox proportional hazards model following Breslow's estimator.
}
