% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breslow_estimator.R
\name{breslow_estimator}
\alias{breslow_estimator}
\title{Breslow estimator of baseline survival}
\usage{
breslow_estimator(W, Delta, HR)
}
\arguments{
\item{W}{Numeric vector of observed covariate values (including censored opens).}

\item{Delta}{Numeric vector of censoring indicators to accompany \code{W}. Note that \code{Delta = 0} is interpreted as a censored observation.}

\item{HR}{Numeric vector of hazard ratios (HR) corresponding to values in \code{W}.}
}
\value{
If \code{x} is not \code{NULL}, then a scalar value for the baseline survival function at value \code{x}. Otherwise, a list with the following three elements is returned:
\item{times}{a vector of the unique observed failure times}
\item{basesurv}{baseline survival estimates}
\item{basehaz}{baseline cumulative hazard estimates}
}
\description{
Estimates the baseline survival function from a Cox proportional hazards model following Breslow's estimator.
}
