% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breslow_estimator.R
\name{breslow_estimator}
\alias{breslow_estimator}
\title{Breslow estimator of baseline survival}
\usage{
breslow_estimator(time, delta, hr, data)
}
\arguments{
\item{time}{String column name for the observed times.}

\item{delta}{String column name for the censoring indicator of the covariate.}

\item{hr}{String column name for the hazard ratios calculated from a \code{coxph} model fit.}

\item{data}{Datafrane containing columns \code{time}, \code{delta}, and \code{hr}.}
}
\value{
A dataframe of the same length and in the same order as \code{data} with the following columns:
\item{times}{unique values of \code{data[, time]}}
\item{basesurv}{baseline survival estimates}
\item{basehaz}{baseline cumulative hazard estimates}
}
\description{
Estimates the baseling survival function from a Cox proportional hazards model following Breslow's esimator.
}
