% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_simulation.R
\name{cox_simulation}
\alias{cox_simulation}
\title{Simulate survival response based on Cox PH model}
\usage{
cox_simulation(
  n,
  logHR,
  covariate,
  dist = "Exponential",
  lambda = 1,
  nu = NULL,
  alpha = NULL
)
}
\arguments{
\item{n}{sample size}

\item{logHR}{coefficients for linear predictor of Cox model}

\item{covariate}{covaraites for linear predictor of Cox model}

\item{dist}{desired outcome distribution; choices include "Exponential," (default) "Weibull," or "Gompertz"}

\item{lambda}{simulation parameter for Exponential, Weibull, and Gompertz distributions}

\item{nu}{simulation parameter for Weibull distributions}

\item{alpha}{simulation parameter for Gompertz distribution}
}
\value{
a response vector of length \code{n} generated by the specified Cox model
}
\description{
Simulate survival response based on Cox PH model that follows either
exponential, Weibull, or Gompertz distribution. Linear predictor for
the Cox model is given by \eqn{covariate * logHR}
}
